Tree Casi Listo

+++ C3DCombiner/Tree/TitusTree.cs
@@ -10,25 +10,35 @@ using Microsoft.VisualBasic.FileIO;
 
 namespace C3DCombiner
 {
-    public List<String> pahs = new List<string>();
 
-    class TitusTree: TreeView
+    class TitusTree : TreeView
     {
+        List<String> Paths;
+
         public TitusTree()
-        {            
+        {
+            this.Paths = new List<string>();
             PathSeparator = @"\";
             this.Dock = DockStyle.Fill;
-            
             setImageIco();
             setMenuContext();
             //setRutaRaiz();
             //Refrescar();
 
             DoubleClick += TitusTree_DoubleClick;
-            NodeMouseClick += TitusTree_Click;            
+            NodeMouseClick += TitusTree_Click;
         }
 
-       
+
+        public void AgregarDirectorio(String directorio)
+        {
+            if (!Paths.Contains(directorio))
+            {
+                Paths.Add(directorio);
+                Refrescar();
+            }
+        }
+
 
         private void TitusTree_Click(object sender, TreeNodeMouseClickEventArgs e)
         {
@@ -42,7 +52,7 @@ namespace C3DCombiner
         {
             int contador = 0;
             int i = raiz.Length - 1;
-            while(contador < 1 && i >= 0)
+            while (contador < 1 && i >= 0)
             {
                 if (raiz[i].ToString() == @"\")
                 {
@@ -52,43 +62,48 @@ namespace C3DCombiner
             }
             //subraiz = raiz.Substring(0, i+1) + @"\";
         }
-        
-        public void Refrescar(String raiz)
+
+        public void Refrescar()
         {
-            if (FileSystem.DirectoryExists(raiz))
+            this.Nodes.Clear();
+            foreach (String path in Paths)
             {
-                ListaDirectorio(this, raiz);
-            }else
-            {
-                this.Nodes.Clear();
-            }   
+                if (FileSystem.DirectoryExists(path))
+                {
+                    ListaDirectorio(this, path);
+                }
+                else
+                {
+                    this.Nodes.Clear();
+                }
+            }
         }
 
         private void ListaDirectorio(TreeView Vista, string ruta)
         {
-            Vista.Nodes.Clear();
             var raizDirectorioInfo = new DirectoryInfo(ruta);
             if (FileSystem.DirectoryExists(ruta))
             {
                 Vista.Nodes.Add(CrearDirectorioNodo(raizDirectorioInfo));
                 Vista.GetNodeAt(0, 0).Expand();
             }
-            
+
         }
         private void setImageIco()
         {
             ImageList lista = new ImageList();
-            lista.Images.Add("Carpeta", Properties.Resources.Nuevo);
-            lista.Images.Add("Archivo", Image.FromFile(@"jc.ico"));
-            lista.Images.Add("Archivo", Image.FromFile(@"acf.ico"));
+            lista.Images.Add("Carpeta", Image.FromFile(@"carpeta.ico"));
+            lista.Images.Add("Archivo", Image.FromFile(@"ocl.ico"));
+            lista.Images.Add("Archivo", Image.FromFile(@"tree.ico"));
+            lista.Images.Add("Archivo", Image.FromFile(@"ddd.ico"));
             this.ImageList = lista;
         }
 
         private void setMenuContext()
         {
-            ContextMenuStrip docMenu = new ContextMenuStrip();            
+            ContextMenuStrip docMenu = new ContextMenuStrip();
 
-            ToolStripMenuItem nuevoacf= new ToolStripMenuItem("Crear archivo", Properties.Resources.Nuevo);
+            ToolStripMenuItem nuevoacf = new ToolStripMenuItem("Crear archivo", Properties.Resources.Nuevo);
             nuevoacf.Click += NuevoArchivo_Click;
 
             ToolStripMenuItem LCrearCarpeta = new ToolStripMenuItem("Crear carpeta");
@@ -100,41 +115,46 @@ namespace C3DCombiner
             ToolStripMenuItem LRuta = new ToolStripMenuItem("Ver ubicacion");
             LRuta.Click += LRuta_Click;
 
-            docMenu.Items.AddRange(new ToolStripMenuItem[] { nuevoacf, LCrearCarpeta,LEliminar, LRuta});
+            docMenu.Items.AddRange(new ToolStripMenuItem[] { nuevoacf, LCrearCarpeta, LEliminar, LRuta });
 
             this.ContextMenuStrip = docMenu;
         }
 
         private void LCrearCarpeta_Click(object sender, EventArgs e)
         {
-            string ruta = /*subraiz +*/ SelectedNode.FullPath;
-            if (FileSystem.DirectoryExists(ruta))
+            if (SelectedNode != null)
             {
-                ruta = /*subraiz + */SelectedNode.FullPath;
-                CrearCarpeta(ruta);
-            }else
-            {
-                ruta = /*subraiz + */SelectedNode.Parent.FullPath;
-                CrearCarpeta(ruta);
+                string ruta = SelectedNode.Name;
+                if (FileSystem.DirectoryExists(ruta))
+                {
+                    ruta = SelectedNode.Name;
+                    CrearCarpeta(ruta);
+                }
+                else
+                {
+                    ruta = SelectedNode.Parent.Name;
+                    CrearCarpeta(ruta);
+                }
             }
         }
 
         private void CrearCarpeta(string ruta)
         {
-            string result=  Microsoft.VisualBasic.Interaction.InputBox("Ingrese el nombre de la carpeta", "Crear carpeta","");
- 
+            string result = Microsoft.VisualBasic.Interaction.InputBox("Ingrese el nombre de la carpeta", "Crear carpeta", "");
+
             if (!String.IsNullOrWhiteSpace(result))
             {
                 string nuevaruta = ruta + @"\" + result;
                 if (FileSystem.DirectoryExists(nuevaruta))
                 {
-                    MessageBox.Show("La carpeta ya existe, ingrese otro nombre.","Alerta",MessageBoxButtons.OK,MessageBoxIcon.Information);
-                }else
+                    MessageBox.Show("La carpeta ya existe, ingrese otro nombre.", "Alerta", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                else
                 {
                     Directory.CreateDirectory(nuevaruta);
-                    Refrescar("");//cambiar ruta
+                    Refrescar();
                 }
-                
+
             }
         }
 
@@ -143,83 +163,97 @@ namespace C3DCombiner
         {
             if (SelectedNode != null)
             {
-                string ruta = /*subraiz + */SelectedNode.FullPath;
+                string ruta = SelectedNode.Name;
                 MessageBox.Show(ruta, "Ubicacion", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }            
+            }
         }
 
         private void BorrarNodo(object sender, EventArgs e)
         {
-            string ruta = /*subraiz + */SelectedNode.FullPath;
-            if (FileSystem.FileExists(ruta))
+            if (SelectedNode != null)
             {
-                FileSystem.DeleteFile(ruta, UIOption.AllDialogs, RecycleOption.SendToRecycleBin, UICancelOption.DoNothing);
-                Refrescar("");
+                string ruta = SelectedNode.Name;
+                if (FileSystem.FileExists(ruta))
+                {
+                    FileSystem.DeleteFile(ruta, UIOption.AllDialogs, RecycleOption.SendToRecycleBin, UICancelOption.DoNothing);
+                    Refrescar();
+                }
+                else if (FileSystem.DirectoryExists(ruta))
+                {
+                    FileSystem.DeleteDirectory(ruta, UIOption.AllDialogs, RecycleOption.SendToRecycleBin, UICancelOption.DoNothing);
+                    Refrescar();
+                }
             }
-            else if (FileSystem.DirectoryExists(ruta))
-            {
-                FileSystem.DeleteDirectory(ruta, UIOption.AllDialogs, RecycleOption.SendToRecycleBin, UICancelOption.DoNothing);
-                Refrescar("");
-            }            
-        }               
+
+        }
 
         private void NuevoArchivo_Click(object sender, EventArgs e)
         {
-            string ruta = /*subraiz + */SelectedNode.FullPath;
-            if (FileSystem.DirectoryExists(ruta))
+            if (SelectedNode != null)
             {
-                CrearArchivo(ruta);
+                string ruta = SelectedNode.Name;
+                if (FileSystem.DirectoryExists(ruta))
+                {
+                    CrearArchivo(ruta);
+                }
+                else
+                {
+                    ruta = SelectedNode.Parent.Name;
+                    CrearArchivo(ruta);
+                }
             }
-            else
-            {
-                ruta = /*subraiz + */SelectedNode.Parent.FullPath;
-                CrearArchivo(ruta);               
-            }            
         }
         private void CrearArchivo(string ruta)
         {
             SaveFileDialog dialogo = new SaveFileDialog();
-            dialogo.Filter = "Tree File|*.tree|OLC File|*.olc|3D File|*.ddd";
+            dialogo.Filter = "Tree File|*.tree|OLC++ File|*.olc|3D File|*.ddd";
             dialogo.Title = "Crear Archivo";
             dialogo.InitialDirectory = ruta;
             dialogo.ShowDialog();
-            
+
             if (dialogo.FileName != "")
             {
                 System.IO.FileStream fs = (System.IO.FileStream)dialogo.OpenFile();
                 fs.Close();
-                Refrescar("");
+                Refrescar();
             }
         }
-
-        private void OpenLabel_Click(object sender, EventArgs e)
-        {
-            Console.WriteLine(this.SelectedNode.FullPath);
-            
-        }        
 
         private static TreeNode CrearDirectorioNodo(DirectoryInfo DirectorioInfo)
         {
             var DirectorioNodo = new TreeNode(DirectorioInfo.Name);
-            
+
             foreach (var directory in DirectorioInfo.GetDirectories())
             {
                 DirectorioNodo.Nodes.Add(CrearDirectorioNodo(directory));
                 DirectorioNodo.ImageIndex = 0;
                 DirectorioNodo.SelectedImageIndex = 0;
+                DirectorioNodo.Name = DirectorioInfo.FullName;
+
             }
             foreach (var file in DirectorioInfo.GetFiles())
             {
-                if (Path.GetExtension(file.ToString()).ToLower() == ".jc")
-                {                   
-                    DirectorioNodo.Nodes.Add(new TreeNode(file.Name,1,1));
-                              
-                }else if (Path.GetExtension(file.ToString()).ToLower() == ".acf")
+                TreeNode nodo = null;
+                if (Path.GetExtension(file.ToString()).ToLower() == ".ocl")
                 {
-                    DirectorioNodo.Nodes.Add(new TreeNode(file.Name, 2, 2));
+                    nodo = new TreeNode(file.Name, 1, 1);
+                }
+                else if (Path.GetExtension(file.ToString()).ToLower() == ".tree")
+                {
+                    nodo = new TreeNode(file.Name, 2, 2);
+                }
+                else if (Path.GetExtension(file.ToString()).ToLower() == ".ddd")
+                {
+                    nodo = new TreeNode(file.Name, 3, 3);
+                }
+
+                if (nodo != null)
+                {
+                    nodo.Name = file.DirectoryName;
+                    DirectorioNodo.Nodes.Add(nodo);
                 }
             }
-            
+
             return DirectorioNodo;
         }
 
@@ -227,9 +261,11 @@ namespace C3DCombiner
 
         private void TitusTree_DoubleClick(object sender, EventArgs e)
         {
-            string ruta = /*subraiz +*/ SelectedNode.FullPath;
+            string ruta = SelectedNode.Name + "\\" + SelectedNode.Text;
             if (FileSystem.FileExists(ruta))
             {
+                
+                TitusTools.Tabs.abrirTab(SelectedNode.Text, ruta, SelectedNode.SelectedImageIndex - 1);
                 //aqui metes el codigo para abrir un archivo XD
                 Console.WriteLine(ruta);
             }
